Displaying Orders & Placing Orders complete.
Displaying orders is now working and also has the option to update an
order as complete. Placing an order also works along with checking if
the customer ID already has an order, and some data type error checking.
I also added some comments in the files I've been working on to make
them easier to read/understand what's going on.